Fixed based on comments
