+++ hazelcast-client/src/main/java/com/hazelcast/client/spi/impl/ClientInvocationFuture.java
@@ -176,9 +176,6 @@ public class ClientInvocationFuture<V> implements ICompletableFuture<V> {
         if (operationResultVerifier != null) {
             value = operationResultVerifier.verify((V) value);
         }
-        if (value instanceof Throwable) {
-            ExceptionUtil.sneakyThrow((Throwable) value);
-        }
         return value;
     }
