+++ hazelcast-client/src/test/java/com/hazelcast/client/cache/ClientJCacheTypeChecksTest.java
@@ -14,19 +14,25 @@
  * limitations under the License.
  */
 
-package com.hazelcast.cache;
+package com.hazelcast.client.cache;
 
+import com.hazelcast.cache.ICache;
 import com.hazelcast.cache.impl.HazelcastServerCachingProvider;
+import com.hazelcast.client.HazelcastClient;
+import com.hazelcast.client.cache.impl.HazelcastClientCachingProvider;
+import com.hazelcast.client.config.ClientConfig;
 import com.hazelcast.config.CacheConfig;
+import com.hazelcast.config.Config;
+import com.hazelcast.config.JoinConfig;
+import com.hazelcast.core.Hazelcast;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.ICompletableFuture;
 import com.hazelcast.test.AssertTask;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.HazelcastTestSupport;
-import com.hazelcast.test.TestHazelcastInstanceFactory;
 import com.hazelcast.test.annotation.QuickTest;
-import org.junit.After;
-import org.junit.Before;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
@@ -34,7 +40,6 @@ import org.junit.runner.RunWith;
 import javax.cache.Cache;
 import javax.cache.CacheManager;
 import javax.cache.configuration.CompleteConfiguration;
-
 import java.lang.reflect.Field;
 
 import static junit.framework.Assert.assertNotNull;
@@ -42,49 +47,62 @@ import static junit.framework.Assert.assertNull;
 
 @RunWith(HazelcastSerialClassRunner.class)
 @Category(QuickTest.class)
-public class JCacheTypeChecks extends HazelcastTestSupport {
+public class ClientJCacheTypeChecksTest
+        extends HazelcastTestSupport {
 
-    private TestHazelcastInstanceFactory factory;
-    private HazelcastInstance hz1;
-    private HazelcastInstance hz2;
+    private static HazelcastInstance hz1;
+    private static HazelcastInstance hz2;
+    private static HazelcastInstance client;
 
-    private HazelcastServerCachingProvider cachingProvider1;
-    private HazelcastServerCachingProvider cachingProvider2;
+    private static HazelcastServerCachingProvider cachingProvider1;
+    private static HazelcastServerCachingProvider cachingProvider2;
+    private static HazelcastClientCachingProvider cachingProvider3;
 
-    @Before
-    public void init() {
-        factory = new TestHazelcastInstanceFactory(2);
-        hz1 = factory.newHazelcastInstance();
-        hz2 = factory.newHazelcastInstance();
+    @BeforeClass
+    public static void init() {
+        Config config = new Config();
+        JoinConfig join = config.getNetworkConfig().getJoin();
+        join.getMulticastConfig().setEnabled(false);
+        join.getTcpIpConfig().setEnabled(true);
+        join.getTcpIpConfig().addMember("127.0.0.1");
+
+        hz1 = Hazelcast.newHazelcastInstance(config);
+        hz2 = Hazelcast.newHazelcastInstance(config);
+
+        ClientConfig clientConfig = new ClientConfig();
+        clientConfig.getNetworkConfig().addAddress("127.0.0.1");
+        client = HazelcastClient.newHazelcastClient(clientConfig);
+
         cachingProvider1 = HazelcastServerCachingProvider.createCachingProvider(hz1);
         cachingProvider2 = HazelcastServerCachingProvider.createCachingProvider(hz2);
+        cachingProvider3 = HazelcastClientCachingProvider.createCachingProvider(client);
     }
 
-    @After
-    public void tear() {
+    @AfterClass
+    public static void tear() {
         cachingProvider1.close();
         cachingProvider2.close();
-        factory.shutdownAll();
+        cachingProvider3.close();
+        HazelcastClient.shutdownAll();
+        Hazelcast.shutdownAll();
     }
 
     @Test(expected = ClassCastException.class)
     public void test_check_types_on_put() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider1.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>()
                 .setTypes(Integer.class, String.class);
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
+                CacheManager cm1 = cachingProvider1.getCacheManager();
+                assertNotNull(cm1.getCache(cacheName, Integer.class, String.class));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName, Integer.class, String.class));
             }
@@ -98,19 +116,17 @@ public class JCacheTypeChecks extends HazelcastTestSupport {
     public void test_check_types_on_get() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider1.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
+                CacheManager cm1 = cachingProvider1.getCacheManager();
+                assertNotNull(cm1.getCache(cacheName));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName));
             }
@@ -119,12 +135,7 @@ public class JCacheTypeChecks extends HazelcastTestSupport {
         Cache rawCache = (Cache) cache;
         rawCache.put(1, 1);
 
-        Class<?> clazz = Class.forName("com.hazelcast.cache.impl.AbstractCacheProxyBase");
-        Field configField = clazz.getDeclaredField("cacheConfig");
-        configField.setAccessible(true);
-
-        CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
+        forceConfigToTypes(rawCache, Integer.class, String.class);
 
         cache.get(1);
     }
@@ -133,19 +144,17 @@ public class JCacheTypeChecks extends HazelcastTestSupport {
     public void test_check_types_on_getandremove() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider1.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
+                CacheManager cm1 = cachingProvider1.getCacheManager();
+                assertNotNull(cm1.getCache(cacheName));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName));
             }
@@ -154,12 +163,7 @@ public class JCacheTypeChecks extends HazelcastTestSupport {
         Cache rawCache = (Cache) cache;
         rawCache.put(1, 1);
 
-        Class<?> clazz = Class.forName("com.hazelcast.cache.impl.AbstractCacheProxyBase");
-        Field configField = clazz.getDeclaredField("cacheConfig");
-        configField.setAccessible(true);
-
-        CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
+        forceConfigToTypes(rawCache, Integer.class, String.class);
 
         cache.getAndRemove(1);
     }
@@ -168,19 +172,17 @@ public class JCacheTypeChecks extends HazelcastTestSupport {
     public void test_check_types_on_replace() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider1.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
+                CacheManager cm1 = cachingProvider1.getCacheManager();
+                assertNotNull(cm1.getCache(cacheName));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName));
             }
@@ -189,69 +191,27 @@ public class JCacheTypeChecks extends HazelcastTestSupport {
         Cache rawCache = (Cache) cache;
         rawCache.put(1, 1);
 
-        Class<?> clazz = Class.forName("com.hazelcast.cache.impl.AbstractCacheProxyBase");
-        Field configField = clazz.getDeclaredField("cacheConfig");
-        configField.setAccessible(true);
-
-        CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
+        forceConfigToTypes(rawCache, Integer.class, String.class);
 
         rawCache.replace(1, 3);
-    }
-
-    @Test(expected = ClassCastException.class)
-    public void test_check_types_on_getandreplace() throws Exception {
-        final String cacheName = randomMapName();
-
-        CacheManager cacheManager = cachingProvider1.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
-
-        CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
-
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
-
-        assertTrueEventually(new AssertTask() {
-            @Override
-            public void run() throws Exception {
-                CacheManager cm2 = cachingProvider2.getCacheManager();
-                assertNotNull(cm2.getCache(cacheName));
-            }
-        });
-
-        Cache rawCache = (Cache) cache;
-        rawCache.put(1, 1);
-
-        Class<?> clazz = Class.forName("com.hazelcast.cache.impl.AbstractCacheProxyBase");
-        Field configField = clazz.getDeclaredField("cacheConfig");
-        configField.setAccessible(true);
-
-        CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
-
-        rawCache.getAndReplace(1, 3);
     }
 
     @Test(expected = ClassCastException.class)
     public void test_check_types_on_put_async() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider1.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>()
                 .setTypes(Integer.class, String.class);
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
+                CacheManager cm1 = cachingProvider1.getCacheManager();
+                assertNotNull(cm1.getCache(cacheName, Integer.class, String.class));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName, Integer.class, String.class));
             }
@@ -266,19 +226,17 @@ public class JCacheTypeChecks extends HazelcastTestSupport {
     public void test_check_types_on_get_async() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider1.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
+                CacheManager cm1 = cachingProvider1.getCacheManager();
+                assertNotNull(cm1.getCache(cacheName));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName));
             }
@@ -287,12 +245,7 @@ public class JCacheTypeChecks extends HazelcastTestSupport {
         Cache rawCache = (Cache) cache;
         rawCache.put(1, 1);
 
-        Class<?> clazz = Class.forName("com.hazelcast.cache.impl.AbstractCacheProxyBase");
-        Field configField = clazz.getDeclaredField("cacheConfig");
-        configField.setAccessible(true);
-
-        CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
+        forceConfigToTypes(rawCache, Integer.class, String.class);
 
         ICompletableFuture future = cache.unwrap(ICache.class).getAsync(1);
         future.get();
@@ -302,19 +255,17 @@ public class JCacheTypeChecks extends HazelcastTestSupport {
     public void test_check_types_on_getandremove_async() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider1.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
+                CacheManager cm1 = cachingProvider1.getCacheManager();
+                assertNotNull(cm1.getCache(cacheName));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName));
             }
@@ -323,12 +274,7 @@ public class JCacheTypeChecks extends HazelcastTestSupport {
         Cache rawCache = (Cache) cache;
         rawCache.put(1, 1);
 
-        Class<?> clazz = Class.forName("com.hazelcast.cache.impl.AbstractCacheProxyBase");
-        Field configField = clazz.getDeclaredField("cacheConfig");
-        configField.setAccessible(true);
-
-        CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
+        forceConfigToTypes(rawCache, Integer.class, String.class);
 
         ICompletableFuture future = cache.unwrap(ICache.class).getAndRemoveAsync(1);
         future.get();
@@ -338,19 +284,17 @@ public class JCacheTypeChecks extends HazelcastTestSupport {
     public void test_check_types_on_replace_async() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider1.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
+                CacheManager cm1 = cachingProvider1.getCacheManager();
+                assertNotNull(cm1.getCache(cacheName));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName));
             }
@@ -359,14 +303,34 @@ public class JCacheTypeChecks extends HazelcastTestSupport {
         ICache rawCache = (ICache) cache.unwrap(ICache.class);
         rawCache.put(1, 1);
 
-        Class<?> clazz = Class.forName("com.hazelcast.cache.impl.AbstractCacheProxyBase");
+        forceConfigToTypes(rawCache, Integer.class, String.class);
+
+        ICompletableFuture future = rawCache.replaceAsync(1, 3);
+        future.get();
+    }
+
+    private Cache<Integer, String> getCache(String cacheName, CacheManager cacheManager,
+                                            CompleteConfiguration<Integer, String> config) {
+        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
+        assertNotNull(cache);
+        return cache;
+    }
+
+    private CacheManager getCacheManager(String cacheName) {
+        CacheManager cacheManager = cachingProvider3.getCacheManager();
+        assertNotNull(cacheManager);
+
+        assertNull(cacheManager.getCache(cacheName));
+        return cacheManager;
+    }
+
+    private void forceConfigToTypes(Cache rawCache, Class<?> keyType, Class<?> valueType)
+            throws ClassNotFoundException, NoSuchFieldException, IllegalAccessException {
+        Class<?> clazz = Class.forName("com.hazelcast.client.cache.impl.AbstractClientCacheProxyBase");
         Field configField = clazz.getDeclaredField("cacheConfig");
         configField.setAccessible(true);
 
         CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
-
-        ICompletableFuture future = rawCache.replaceAsync(1, 3);
-        future.get();
+        cacheConfig.setTypes(keyType, valueType);
     }
 }
