+++ hazelcast/src/test/java/com/hazelcast/cache/JCacheTypeChecksTest.java
@@ -14,25 +14,19 @@
  * limitations under the License.
  */
 
-package com.hazelcast.client.cache;
+package com.hazelcast.cache;
 
-import com.hazelcast.cache.ICache;
 import com.hazelcast.cache.impl.HazelcastServerCachingProvider;
-import com.hazelcast.client.HazelcastClient;
-import com.hazelcast.client.cache.impl.HazelcastClientCachingProvider;
-import com.hazelcast.client.config.ClientConfig;
 import com.hazelcast.config.CacheConfig;
-import com.hazelcast.config.Config;
-import com.hazelcast.config.JoinConfig;
-import com.hazelcast.core.Hazelcast;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.ICompletableFuture;
 import com.hazelcast.test.AssertTask;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.HazelcastTestSupport;
+import com.hazelcast.test.TestHazelcastInstanceFactory;
 import com.hazelcast.test.annotation.QuickTest;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
@@ -40,6 +34,7 @@ import org.junit.runner.RunWith;
 import javax.cache.Cache;
 import javax.cache.CacheManager;
 import javax.cache.configuration.CompleteConfiguration;
+
 import java.lang.reflect.Field;
 
 import static junit.framework.Assert.assertNotNull;
@@ -47,66 +42,47 @@ import static junit.framework.Assert.assertNull;
 
 @RunWith(HazelcastSerialClassRunner.class)
 @Category(QuickTest.class)
-public class ClientJCacheTypeChecks
+public class JCacheTypeChecksTest
         extends HazelcastTestSupport {
 
-    private static HazelcastInstance hz1;
-    private static HazelcastInstance hz2;
-    private static HazelcastInstance client;
+    private TestHazelcastInstanceFactory factory;
+    private HazelcastInstance hz1;
+    private HazelcastInstance hz2;
 
-    private static HazelcastServerCachingProvider cachingProvider1;
-    private static HazelcastServerCachingProvider cachingProvider2;
-    private static HazelcastClientCachingProvider cachingProvider3;
+    private HazelcastServerCachingProvider cachingProvider1;
+    private HazelcastServerCachingProvider cachingProvider2;
 
-    @BeforeClass
-    public static void init() {
-        Config config = new Config();
-        JoinConfig join = config.getNetworkConfig().getJoin();
-        join.getMulticastConfig().setEnabled(false);
-        join.getTcpIpConfig().setEnabled(true);
-        join.getTcpIpConfig().addMember("127.0.0.1");
-
-        hz1 = Hazelcast.newHazelcastInstance(config);
-        hz2 = Hazelcast.newHazelcastInstance(config);
-
-        ClientConfig clientConfig = new ClientConfig();
-        clientConfig.getNetworkConfig().addAddress("127.0.0.1");
-        client = HazelcastClient.newHazelcastClient(clientConfig);
-
+    @Before
+    public void init() {
+        factory = new TestHazelcastInstanceFactory(2);
+        hz1 = factory.newHazelcastInstance();
+        hz2 = factory.newHazelcastInstance();
         cachingProvider1 = HazelcastServerCachingProvider.createCachingProvider(hz1);
         cachingProvider2 = HazelcastServerCachingProvider.createCachingProvider(hz2);
-        cachingProvider3 = HazelcastClientCachingProvider.createCachingProvider(client);
     }
 
-    @AfterClass
-    public static void tear() {
+    @After
+    public void tear() {
         cachingProvider1.close();
         cachingProvider2.close();
-        cachingProvider3.close();
-        HazelcastClient.shutdownAll();
-        Hazelcast.shutdownAll();
+        factory.shutdownAll();
     }
 
     @Test(expected = ClassCastException.class)
     public void test_check_types_on_put() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider3.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>()
                 .setTypes(Integer.class, String.class);
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
-            public void run() throws Exception {
-                CacheManager cm1 = cachingProvider1.getCacheManager();
-                assertNotNull(cm1.getCache(cacheName, Integer.class, String.class));
+            public void run()
+                    throws Exception {
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName, Integer.class, String.class));
             }
@@ -120,21 +96,15 @@ public class ClientJCacheTypeChecks
     public void test_check_types_on_get() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider3.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
-                CacheManager cm1 = cachingProvider1.getCacheManager();
-                assertNotNull(cm1.getCache(cacheName));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName));
             }
@@ -143,12 +113,7 @@ public class ClientJCacheTypeChecks
         Cache rawCache = (Cache) cache;
         rawCache.put(1, 1);
 
-        Class<?> clazz = Class.forName("com.hazelcast.client.cache.impl.AbstractClientCacheProxyBase");
-        Field configField = clazz.getDeclaredField("cacheConfig");
-        configField.setAccessible(true);
-
-        CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
+        forceConfigToTypes(rawCache, Integer.class, String.class);
 
         cache.get(1);
     }
@@ -157,21 +122,15 @@ public class ClientJCacheTypeChecks
     public void test_check_types_on_getandremove() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider3.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
-                CacheManager cm1 = cachingProvider1.getCacheManager();
-                assertNotNull(cm1.getCache(cacheName));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName));
             }
@@ -180,12 +139,7 @@ public class ClientJCacheTypeChecks
         Cache rawCache = (Cache) cache;
         rawCache.put(1, 1);
 
-        Class<?> clazz = Class.forName("com.hazelcast.client.cache.impl.AbstractClientCacheProxyBase");
-        Field configField = clazz.getDeclaredField("cacheConfig");
-        configField.setAccessible(true);
-
-        CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
+        forceConfigToTypes(rawCache, Integer.class, String.class);
 
         cache.getAndRemove(1);
     }
@@ -194,21 +148,15 @@ public class ClientJCacheTypeChecks
     public void test_check_types_on_replace() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider3.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
-                CacheManager cm1 = cachingProvider1.getCacheManager();
-                assertNotNull(cm1.getCache(cacheName));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName));
             }
@@ -217,36 +165,51 @@ public class ClientJCacheTypeChecks
         Cache rawCache = (Cache) cache;
         rawCache.put(1, 1);
 
-        Class<?> clazz = Class.forName("com.hazelcast.client.cache.impl.AbstractClientCacheProxyBase");
-        Field configField = clazz.getDeclaredField("cacheConfig");
-        configField.setAccessible(true);
-
-        CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
+        forceConfigToTypes(rawCache, Integer.class, String.class);
 
         rawCache.replace(1, 3);
+    }
+
+    @Test(expected = ClassCastException.class)
+    public void test_check_types_on_getandreplace() throws Exception {
+        final String cacheName = randomMapName();
+
+        CacheManager cacheManager = getCacheManager(cacheName);
+
+        CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
+
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
+
+        assertTrueEventually(new AssertTask() {
+            @Override
+            public void run() throws Exception {
+                CacheManager cm2 = cachingProvider2.getCacheManager();
+                assertNotNull(cm2.getCache(cacheName));
+            }
+        });
+
+        Cache rawCache = (Cache) cache;
+        rawCache.put(1, 1);
+
+        forceConfigToTypes(rawCache, Integer.class, String.class);
+
+        rawCache.getAndReplace(1, 3);
     }
 
     @Test(expected = ClassCastException.class)
     public void test_check_types_on_put_async() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider3.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>()
                 .setTypes(Integer.class, String.class);
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
-                CacheManager cm1 = cachingProvider1.getCacheManager();
-                assertNotNull(cm1.getCache(cacheName, Integer.class, String.class));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName, Integer.class, String.class));
             }
@@ -261,21 +224,15 @@ public class ClientJCacheTypeChecks
     public void test_check_types_on_get_async() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider3.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
-                CacheManager cm1 = cachingProvider1.getCacheManager();
-                assertNotNull(cm1.getCache(cacheName));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName));
             }
@@ -284,12 +241,7 @@ public class ClientJCacheTypeChecks
         Cache rawCache = (Cache) cache;
         rawCache.put(1, 1);
 
-        Class<?> clazz = Class.forName("com.hazelcast.client.cache.impl.AbstractClientCacheProxyBase");
-        Field configField = clazz.getDeclaredField("cacheConfig");
-        configField.setAccessible(true);
-
-        CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
+        forceConfigToTypes(rawCache, Integer.class, String.class);
 
         ICompletableFuture future = cache.unwrap(ICache.class).getAsync(1);
         future.get();
@@ -299,21 +251,15 @@ public class ClientJCacheTypeChecks
     public void test_check_types_on_getandremove_async() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider3.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
-                CacheManager cm1 = cachingProvider1.getCacheManager();
-                assertNotNull(cm1.getCache(cacheName));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName));
             }
@@ -322,12 +268,7 @@ public class ClientJCacheTypeChecks
         Cache rawCache = (Cache) cache;
         rawCache.put(1, 1);
 
-        Class<?> clazz = Class.forName("com.hazelcast.client.cache.impl.AbstractClientCacheProxyBase");
-        Field configField = clazz.getDeclaredField("cacheConfig");
-        configField.setAccessible(true);
-
-        CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
+        forceConfigToTypes(rawCache, Integer.class, String.class);
 
         ICompletableFuture future = cache.unwrap(ICache.class).getAndRemoveAsync(1);
         future.get();
@@ -337,21 +278,15 @@ public class ClientJCacheTypeChecks
     public void test_check_types_on_replace_async() throws Exception {
         final String cacheName = randomMapName();
 
-        CacheManager cacheManager = cachingProvider3.getCacheManager();
-        assertNotNull(cacheManager);
-
-        assertNull(cacheManager.getCache(cacheName));
+        CacheManager cacheManager = getCacheManager(cacheName);
 
         CompleteConfiguration<Integer, String> config = new CacheConfig<Integer, String>();
 
-        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
-        assertNotNull(cache);
+        Cache<Integer, String> cache = getCache(cacheName, cacheManager, config);
 
         assertTrueEventually(new AssertTask() {
             @Override
             public void run() throws Exception {
-                CacheManager cm1 = cachingProvider1.getCacheManager();
-                assertNotNull(cm1.getCache(cacheName));
                 CacheManager cm2 = cachingProvider2.getCacheManager();
                 assertNotNull(cm2.getCache(cacheName));
             }
@@ -360,14 +295,34 @@ public class ClientJCacheTypeChecks
         ICache rawCache = (ICache) cache.unwrap(ICache.class);
         rawCache.put(1, 1);
 
-        Class<?> clazz = Class.forName("com.hazelcast.client.cache.impl.AbstractClientCacheProxyBase");
+        forceConfigToTypes(rawCache, Integer.class, String.class);
+
+        ICompletableFuture future = rawCache.replaceAsync(1, 3);
+        future.get();
+    }
+
+    private Cache<Integer, String> getCache(String cacheName, CacheManager cacheManager,
+                                            CompleteConfiguration<Integer, String> config) {
+        Cache<Integer, String> cache = cacheManager.createCache(cacheName, config);
+        assertNotNull(cache);
+        return cache;
+    }
+
+    private CacheManager getCacheManager(String cacheName) {
+        CacheManager cacheManager = cachingProvider1.getCacheManager();
+        assertNotNull(cacheManager);
+
+        assertNull(cacheManager.getCache(cacheName));
+        return cacheManager;
+    }
+
+    private void forceConfigToTypes(Cache rawCache, Class<?> keyType, Class<?> valueType)
+            throws ClassNotFoundException, NoSuchFieldException, IllegalAccessException {
+        Class<?> clazz = Class.forName("com.hazelcast.cache.impl.AbstractCacheProxyBase");
         Field configField = clazz.getDeclaredField("cacheConfig");
         configField.setAccessible(true);
 
         CacheConfig cacheConfig = (CacheConfig) configField.get(rawCache);
-        cacheConfig.setTypes(Integer.class, String.class);
-
-        ICompletableFuture future = rawCache.replaceAsync(1, 3);
-        future.get();
+        cacheConfig.setTypes(keyType, valueType);
     }
 }
